Move templates behind a feature.
