+++ crates/cli/Cargo.toml
@@ -27,7 +27,7 @@ moon_bun_tool = { path = "../bun/tool" }
 moon_cache = { path = "../../nextgen/cache" }
 moon_codegen = { path = "../../nextgen/codegen" }
 moon_common = { path = "../../nextgen/common" }
-moon_config = { path = "../../nextgen/config" }
+moon_config = { path = "../../nextgen/config", features = ["template"] }
 moon_env = { path = "../../nextgen/env" }
 moon_lang = { path = "../core/lang" }
 moon_node_lang = { path = "../node/lang" }
