+++ nextgen/codegen/Cargo.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/moonrepo/moon"
 
 [dependencies]
 moon_common = { path = "../common" }
-moon_config = { path = "../config", features = ["loader"] }
+moon_config = { path = "../config", features = ["loader", "template"] }
 content_inspector = "0.2.4"
 convert_case = "0.6.0"
 miette = { workspace = true }
