+++ nextgen/config/src/lib.rs
@@ -14,6 +14,9 @@ mod validate;
 mod workspace;
 mod workspace_config;
 
+#[cfg(feature = "template")]
+mod templates;
+
 pub use inherited_tasks_config::*;
 pub use language_platform::*;
 pub use portable_path::*;
@@ -31,38 +34,5 @@ pub use version_spec::{UnresolvedVersionSpec, VersionSpec};
 pub use workspace::*;
 pub use workspace_config::*;
 
-pub fn load_workspace_config_template() -> &'static str {
-    include_str!("../templates/workspace.yml")
-}
-
-pub fn load_toolchain_config_template() -> &'static str {
-    include_str!("../templates/toolchain.yml")
-}
-
-pub fn load_toolchain_bun_config_template() -> &'static str {
-    include_str!("../templates/toolchain_bun.yml")
-}
-
-pub fn load_toolchain_deno_config_template() -> &'static str {
-    include_str!("../templates/toolchain_deno.yml")
-}
-
-pub fn load_toolchain_node_config_template() -> &'static str {
-    include_str!("../templates/toolchain_node.yml")
-}
-
-pub fn load_toolchain_rust_config_template() -> &'static str {
-    include_str!("../templates/toolchain_rust.yml")
-}
-
-pub fn load_toolchain_typescript_config_template() -> &'static str {
-    include_str!("../templates/toolchain_typescript.yml")
-}
-
-pub fn load_tasks_config_template() -> &'static str {
-    include_str!("../templates/tasks.yml")
-}
-
-pub fn load_template_config_template() -> &'static str {
-    include_str!("../templates/template.yml")
-}
+#[cfg(feature = "template")]
+pub use templates::*;
